+++ src/Utilities/Container.h
@@ -1,0 +1,24 @@
+// Copyright (c) 2019, The TurtleCoin Developers
+// 
+// Please see the included LICENSE file for more information.
+
+#include <algorithm>
+
+#include <vector>
+
+namespace Utilities
+{
+
+template <typename T, typename Function>
+std::vector<T> filter(const std::vector<T> &input, Function predicate)
+{
+    std::vector<T> result;
+
+    std::copy_if(
+        input.begin(), input.end(), std::back_inserter(result), predicate
+    );
+
+    return result;
+}
+
+} // namespace Utilities
