+++ src/zedwallet++/Utilities.h
@@ -23,16 +23,4 @@ void confirmPassword(
 
 uint64_t getScanHeight();
 
-template <typename T, typename Function>
-std::vector<T> filter(const std::vector<T> &input, Function predicate)
-{
-    std::vector<T> result;
-
-    std::copy_if(
-        input.begin(), input.end(), std::back_inserter(result), predicate
-    );
-
-    return result;
-}
-
 } // namespace
